# services/gateway/swagger.yaml
openapi: 3.0.0
info:
  title: AI Code Reviewer API
  description: |
    API complète pour l'analyse automatique de code source via Intelligence Artificielle.
    
    **Architecture**: Microservices avec API Gateway
    
    **Services**:
    - Auth Service (Port 3001) - Authentification JWT
    - Code Service (Port 3002) - Gestion des analyses
    - AI Service (Port 3003) - Analyse OpenAI
    - Payment Service (Port 3004) - Paiements Stripe
    
    **Authentication**: Utiliser le header `Authorization: Bearer <token>`
  version: 1.0.0
  contact:
    name: Support AI Code Reviewer
    email: support@aicodereviewer.com
servers:
  - url: http://localhost:3000/api
    description: Développement local
  - url: https://api.aicodereviewer.com/api
    description: Production
...
paths:
...
  /auth/register:
    post:
      summary: Créer un nouveau compte
      tags: [Authentification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password, firstName, lastName]
              properties:
                email:
                  type: string
                  format: email
                  example: "oumaima.boullam@example.com"
                password:
                  type: string
                  minLength: 8
                  example: "SecurePass123"
                  description: "Min 8 caractères, 1 majuscule, 1 minuscule, 1 chiffre"
                firstName:
                  type: string
                  example: "Jean"
                lastName:
                  type: string
                  example: "Dupont"
      responses:
        201:
          description: Utilisateur créé avec succès
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        400:
          description: Validation échouée ou email déjà utilisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/login:
    post:
      summary: Se connecter
      tags: [Authentification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email:
                  type: string
                  format: email
                  example: "oumaima.boullam@example.com"
                password:
                  type: string
                  example: "SecurePass123"
      responses:
        200:
          description: Connexion réussie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        401:
          description: Identifiants incorrects
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      summary: Renouveler l'access token
      tags: [Authentification]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken:
                  type: string
                  description: Refresh token obtenu lors de la connexion
      responses:
        200:
          description: Tokens renouvelés
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  tokens:
                    $ref: '#/components/schemas/TokenPair'
  /auth/logout:
    post:
      summary: Se déconnecter
      tags: [Authentification]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
      responses:
        200:
          description: Déconnexion réussie
  /auth/me:
    get:
      summary: Récupérer les informations de l'utilisateur connecté
      tags: [Authentification]
      security:
        - BearerAuth: []
      responses:
        200:
          description: Informations utilisateur
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  user:
                    $ref: '#/components/schemas/User'
  /auth/profile:
    put:
      summary: Mettre à jour le profil
      tags: [Authentification]
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                firstName:
                  type: string
                lastName:
                  type: string
                preferences:
                  type: object
      responses:
        200:
          description: Profil mis à jour

  # ==================== CODE ====================
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT obtenu lors de la connexion
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        isPremium:
          type: boolean
        creditsRemaining:
          type: integer
        premiumEndDate:
          type: string
          format: date-time
          nullable: true
        preferences:
          type: object
          properties:
            emailNotifications:
              type: boolean
            darkMode:
              type: boolean
            language:
              type: string
    TokenPair:
      type: object
      properties:
        accessToken:
          type: string
          description: JWT valide 15 minutes
        refreshToken:
          type: string
          description: JWT valide 7 jours
        expiresIn:
          type: integer
          description: Secondes avant expiration de l'access token
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        user:
          $ref: '#/components/schemas/User'
        tokens:
          $ref: '#/components/schemas/TokenPair'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: string
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
...